[FEATURE] Remind others, not only me
